% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalEu.r
\name{capitalEu}
\alias{capitalEu}
\title{Get European Country Capial City}
\usage{
capitalEu(country, displayWay)
}
\arguments{
\item{country}{country}

\item{displayWay}{way of displaying results}
}
\value{
capital city of given
}
\description{
capitalEu provides functionality which enables to get capital city of some european countries
}
\examples{
capitalEu('Spain', 'mix')

}
