% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDirectory.R
\name{cleanDirectory}
\alias{cleanDirectory}
\title{Drop All Csv From Given Directory}
\usage{
cleanDirectory(path)
}
\arguments{
\item{path}{directory where csv to delete are stored, as a string.
Path should be ended without slash (/).}
}
\value{
Directory without csv files, console output presents names of dropped csv's.
}
\description{
Deletes all csv files from given directory. Use with caution.
}
