% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTodayCurrencyToEur.R
\name{getTodayCurrencyToEur}
\alias{getTodayCurrencyToEur}
\title{Get Daily Euro Exchange Rate}
\usage{
getTodayCurrencyToEur(result = "table")
}
\arguments{
\item{result}{specifies output type, table or status}
}
\value{
returns table or status, depending on given parameter
}
\description{
getTodayCurrencyToEur is used to get today's exchange rate for Euro
}
\examples{
getTodayCurrencyToEur()

}
