% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotM.R
\name{plotM}
\alias{plotM}
\title{Simple Plot Manipulation}
\usage{
plotM(data, col1, col2)
}
\arguments{
\item{data}{dataframe to plot}

\item{col1}{x column name}

\item{col2}{y column name}
}
\value{
Plot which can be manipulated
}
\description{
plotM can be used to manipulate ranges, styles and
   colors of displayed data
}
