% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearModel.R
\name{linearModel}
\alias{linearModel}
\title{Draws Coefficients Barplot}
\usage{
linearModel(formula, data)
}
\arguments{
\item{formula}{Linear Model Formula}

\item{data}{Data To Model}
}
\value{
Barplot With Linear Model Coefficients Without Intercept
}
\description{
linearModel is used to vizualize coefficients of linear models as barplot.
    It can be used for any linear model equation.
}
\examples{
data = data.frame(y = c(1:100), x1 = runif(100), x2 = c(2:101)+runif(100))
linearModel(y~ x1+x2, data = data)
}
